requirements.txt
discord.py[voice]
python-dotenv
PyNaCl
yt-dlp
pytest
pytest-asyncio
pytest-mock
aiohttp